fix(server): use of internals
assert error
update package.json
