+++ index.js
@@ -1,6 +1,1 @@
-'use strict';
-var server = require('./lib/server');
-
-server.start(function() {
-    console.log('Server started at', server.info.uri);
-});
+module.exports = require('./lib');
