+++ package.json
@@ -1,90 +1,115 @@
 {
-    "name": "hueniversity",
-    "version": "0.0.1",
-    "description": "This is a community learning experiment. The idea is simple - use GitHub as a platform for teaching people coding skills as a group, where everyone is both a student and a teacher. The goal is to learn how to operate such a distributed classroom and then apply that pattern to other topics by other people.",
-    "main": "index.js",
-    "scripts": {
-        "start": "node .",
-        "watch": "pm2 start index.js --watch --name hueniversity",
-        "stop": "pm2 stop --watch hueniversity",
-        "test": "lab -t 100 -L",
-        "test-cov-html": "lab -r html -o coverage.html"
-    },
-    "repository": {
-        "type": "git",
-        "url": "https://github.com/MrBri/hueniversity.git"
-    },
-    "author": "Tyler Briles",
-    "license": "ISC",
-    "bugs": {
-        "url": "https://github.com/MrBri/hueniversity/issues"
-    },
-    "homepage": "https://github.com/MrBri/hueniversity",
-    "dependencies": {
-        "hapi": "^8.4.0"
-    },
-    "devDependencies": {
-        "code": "^1.3.0",
-        "lab": "^5.5.0",
-        "pm2": "^0.12.7"
-    },
-    "eslintConfig": {
-        "rules": {
-            "consistent-return": 0,
-            "vars-on-top": 0,
-            "one-var": 0,
-            "strict": 0,
-            "new-cap": 0,
-            "no-console": 0,
-            "no-constant-condition": 0,
-            "no-empty": 0,
-            "no-native-reassign": 0,
-            "no-underscore-dangle": 0,
-            "no-undef": 0,
-            "no-process-exit": 0,
-            "no-unused-vars": 0,
-            "no-use-before-define": 0,
-            "no-unused-expressions": 0,
-            "no-regex-spaces": 0,
-            "no-shadow": 0,
-            "global-strict": 0,
-            "handle-callback-err": 0,
-            "no-lonely-if": 0,
-            "space-in-brackets": 0,
-            "brace-style": 0,
-            "dot-notation": 1,
-            "eol-last": 1,
-            "camelcase": 1,
-            "no-trailing-spaces": 1,
-            "no-eq-null": 1,
-            "no-extend-native": 1,
-            "no-redeclare": 1,
-            "no-loop-func": 1,
-            "yoda": [1, "never"],
-            "sort-vars": 1,
-            "quotes": [2, "single"],
-            "consistent-this": [2, "self"],
-            "func-style": [2, "expression"],
-            "new-parens": 2,
-            "no-array-constructor": 2,
-            "no-new-object": 2,
-            "no-spaced-func": 2,
-            "no-space-before-semi": 2,
-            "no-mixed-spaces-and-tabs": 2,
-            "space-after-keywords": 2,
-            "semi": [2, "always"],
-            "space-infix-ops": 2,
-            "space-return-throw-case": 2,
-            "space-unary-ops": [1, { "words": true, "nonwords": false }],
-            "eqeqeq": 2,
-            "curly": [2, "all"],
-            "no-eval": 2,
-            "no-else-return": 2,
-            "no-return-assign": 2,
-            "no-new-wrappers": 2,
-            "no-comma-dangle": 2,
-            "no-sparse-arrays": 2,
-            "no-ex-assign": 2
+  "name": "hueniversity",
+  "version": "0.0.1",
+  "description": "This is a community learning experiment. The idea is simple - use GitHub as a platform for teaching people coding skills as a group, where everyone is both a student and a teacher. The goal is to learn how to operate such a distributed classroom and then apply that pattern to other topics by other people.",
+  "main": "index.js",
+  "scripts": {
+    "start": "node .",
+    "watch": "pm2 start index.js --watch --name hueniversity",
+    "stop": "pm2 stop --watch hueniversity",
+    "test": "lab -t 100 -L",
+    "test-cov-html": "lab -r html -o coverage.html"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/MrBri/hueniversity.git"
+  },
+  "author": "Tyler Briles",
+  "license": "BSD",
+  "bugs": {
+    "url": "https://github.com/hueniverse/hueniversity.git"
+  },
+  "homepage": "https://github.com/MrBri/hueniversity",
+  "dependencies": {
+    "hapi": "^8.4.0",
+    "hoek": "^2.11.1"
+  },
+  "devDependencies": {
+    "code": "^1.3.0",
+    "lab": "^5.5.0",
+    "pm2": "^0.12.7"
+  },
+  "eslintConfig": {
+    "rules": {
+      "consistent-return": 0,
+      "vars-on-top": 0,
+      "one-var": 0,
+      "strict": 0,
+      "new-cap": 0,
+      "no-console": 0,
+      "no-constant-condition": 0,
+      "no-empty": 0,
+      "no-native-reassign": 0,
+      "no-underscore-dangle": 0,
+      "no-undef": 0,
+      "no-process-exit": 0,
+      "no-unused-vars": 0,
+      "no-use-before-define": 0,
+      "no-unused-expressions": 0,
+      "no-regex-spaces": 0,
+      "no-shadow": 0,
+      "global-strict": 0,
+      "handle-callback-err": 0,
+      "no-lonely-if": 0,
+      "space-in-brackets": 0,
+      "brace-style": 0,
+      "dot-notation": 1,
+      "eol-last": 1,
+      "camelcase": 1,
+      "no-trailing-spaces": 1,
+      "no-eq-null": 1,
+      "no-extend-native": 1,
+      "no-redeclare": 1,
+      "no-loop-func": 1,
+      "yoda": [
+        1,
+        "never"
+      ],
+      "sort-vars": 1,
+      "quotes": [
+        2,
+        "single"
+      ],
+      "consistent-this": [
+        2,
+        "self"
+      ],
+      "func-style": [
+        2,
+        "expression"
+      ],
+      "new-parens": 2,
+      "no-array-constructor": 2,
+      "no-new-object": 2,
+      "no-spaced-func": 2,
+      "no-space-before-semi": 2,
+      "no-mixed-spaces-and-tabs": 2,
+      "space-after-keywords": 2,
+      "semi": [
+        2,
+        "always"
+      ],
+      "space-infix-ops": 2,
+      "space-return-throw-case": 2,
+      "space-unary-ops": [
+        1,
+        {
+          "words": true,
+          "nonwords": false
         }
+      ],
+      "eqeqeq": 2,
+      "curly": [
+        2,
+        "all"
+      ],
+      "no-eval": 2,
+      "no-else-return": 2,
+      "no-return-assign": 2,
+      "no-new-wrappers": 2,
+      "no-comma-dangle": 2,
+      "no-sparse-arrays": 2,
+      "no-ex-assign": 2
     }
+  }
 }
